Fix portals duplicating the same code 5 times
